I think the recursion method works
Intentionally left the debugging marker so that I can come
back and look how the code is operating.

I initially struggled at this assignment first because I didn't
follow the youtube tutorial posted by the Odin Project. Then
I struggled some more because I hastily copy and pasted a repeated
code snippet from one block to the other and forgot to change the
variable names. Then I struggled some more because I did not know
'each' cannot handle original array manipulation. In my code,
I wanted the elements of 'right' array to be moved into 'bar' array.
It so happens that when I do move an element, the index count does not
update in the next iteration, leading it to never reach the latter
elements. That is, 'each' loop continues based on the original index
count, not the updated array. Chatgpt was barely any help this time.
I am beginning to think that chatgpt is only helpful when the code
is very straightforward.